Some code cleanup & implement error DTO
