+++ reposilite-backend/pom.xml
@@ -138,7 +138,6 @@
         </dependency>
 
         <!-- Tests -->
-        <!-- https://mvnrepository.com/artifact/org.mongodb/mongo-java-driver -->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-engine</artifactId>
