+++ reposilite-backend/src/main/java/org/panda_lang/reposilite/error/ErrorDto.java
@@ -1,0 +1,21 @@
+package org.panda_lang.reposilite.error;
+
+public class ErrorDto {
+
+    private int code;
+    private String message;
+
+    public ErrorDto(int code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    public int getCode() {
+        return this.code;
+    }
+
+    public String getMessage() {
+        return this.message;
+    }
+
+}
